Move cWGANTrainer class to cWGAN.py
